# prototype/1. Technical_ EU STR - Code Single Digital Entry Point/str-ap-internal-0.0.49 (update February 2025)/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A gateway for the electronic transmission of data between online short-term rental platforms and competent authorities, ensuring timely, reliable and efficient data sharing processes\nPriority of development is: 1. listings, 2. orders, 3. activity, 4. area\n\u003cbr\u003e\nTo obtain API credentials, please contact: wouter.travers@pwc.com via e-mail",
        "title": "EU STR - Single Digital Entry Point",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Wouter Travers",
            "email": "wouter.travers@pwc.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/ca/listings": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Retrieve listings submitted to the SDEP",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Retrieve listings submitted to the SDEP",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "limit number of records returned",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.SingleListingData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid limit parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/str/listings": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "submit listing(s) to SDEP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "str"
                ],
                "summary": "submit listing(s) to SDEP",
                "parameters": [
                    {
                        "description": "json formatted ListingsData",
                        "name": "ListingsData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ListingsData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.Status"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing mandatory parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Invalid data field values",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPError"
                        }
                    }
                }
            }
        },
        "/str/number-valid-check/{registrationNumber}": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Registration Number Validation check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "str"
                ],
                "summary": "Registration Number Validation check",
                "parameters": [
                    {
                        "description": "Registration Number",
                        "name": "registrationNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/handler.Numbervalid"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing registration number",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "422": {
                        "description": "Invalid registration number format",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "City name",
                    "type": "string",
                    "example": "Diegem"
                },
                "country": {
                    "description": "Country code with ISO 3166-1 alpha-3 validation",
                    "type": "string",
                    "example": "BEL"
                },
                "postalCode": {
                    "description": "Postal code",
                    "type": "string",
                    "example": "1831"
                },
                "street": {
                    "description": "Street address",
                    "type": "string",
                    "example": "Culliganlaan 5"
                }
            }
        },
        "handler.BadRequestError": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Error message for bad requests",
                    "type": "string",
                    "example": "Wrong data format!"
                }
            }
        },
...
        "handler.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "handler.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message for internal server errors",
                    "type": "string",
                    "example": "An unexpected error occurred"
                }
            }
        },
        "handler.ListingsData": {
            "type": "object",
            "required": [
                "data",
                "metadata"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.RegistrationData"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/handler.Metadata"
                }
            }
        },
        "handler.Metadata": {
            "type": "object",
            "properties": {
                "additionalProp1": {
                    "description": "Placeholder for additional properties",
                    "type": "object"
                },
                "platform": {
                    "description": "Platform name",
                    "type": "string",
                    "example": "booking.com"
                },
                "submissionDate": {
                    "description": "Date of submission",
                    "type": "string",
                    "example": "2024-07-21T17:32:28Z"
                }
            }
        },
...
        "handler.Numbervalid": {
            "type": "object",
            "properties": {
                "isValid": {
                    "description": "Error message for a valid registration number",
                    "type": "string",
                    "example": "True"
                }
            }
        },
        "handler.RegistrationData": {
            "type": "object",
            "properties": {
                "Unit": {
                    "$ref": "#/definitions/handler.Unit"
                },
                "competentAuthorityId_area": {
                    "type": "string",
                    "example": "competentAuthorityId_area"
                },
                "competentAuthorityName_area": {
                    "type": "string",
                    "example": "competentAuthorityName_area"
                },
                "registrationNumber": {
                    "type": "string"
                }
            }
        },
...
        "handler.SingleListingData": {
            "type": "object",
            "required": [
                "data",
                "metadata"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.RegistrationData"
                },
                "metadata": {
                    "$ref": "#/definitions/handler.Metadata"
                }
            }
        },
        "handler.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Status message",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "handler.UnauthorizedError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Error message for unauthorized access",
                    "type": "string",
                    "example": "JWT is invalid"
                }
            }
        },
        "handler.Unit": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/handler.Address"
                },
                "description": {
                    "type": "string"
                },
                "floorLevel": {
                    "type": "string"
                },
                "numberOfRooms": {
                    "type": "integer"
                },
                "obtainedAuth": {
                    "type": "boolean"
                },
                "occupancy": {
                    "type": "integer"
                },
                "purpose": {
                    "type": "string"
                },
                "subjectToAuth": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://tt-dp-dev.eu.auth0.com/authorize?audience=https://str.eu",
            "tokenUrl": "https://tt-dp-dev.eu.auth0.com/oauth/token",
            "scopes": {
                "str": "Grants read access"
            }
        }
    },
...
}